+++ eshop/src/main/java/az/orient/eshop/service/impl/ProductServiceImpl.java
@@ -14,6 +14,8 @@ import az.orient.eshop.service.ProductService;
 import az.orient.eshop.util.Utility;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 import java.util.*;
 
 @Service
@@ -218,7 +220,7 @@ public class ProductServiceImpl implements ProductService {
             Long colorId = reqProductDetails.getColorId();
             Integer stock = reqProductDetails.getStock();
             Currency currency = Currency.fromValue(reqProductDetails.getCurrency().getValue());
-            Float price = reqProductDetails.getPrice();
+            BigDecimal price = reqProductDetails.getPrice();
             if (sizeId == null || stock == null || colorId == null || price == null) {
                 throw new EshopException(ExceptionConstants.INVALID_REQUEST_DATA, "Invalid request data");
             }
@@ -258,7 +260,7 @@ public class ProductServiceImpl implements ProductService {
             Long colorId = reqProductDetails.getColorId();
             Integer stock = reqProductDetails.getStock();
             Currency currency = Currency.fromValue(reqProductDetails.getCurrency().getValue());
-            Float price = reqProductDetails.getPrice();
+            BigDecimal price = reqProductDetails.getPrice();
             if (sizeId == null || stock == null || colorId == null || price == null) {
                 throw new EshopException(ExceptionConstants.INVALID_REQUEST_DATA, "Invaild request data");
             }
